Renamed JDOM template processor to main processor and adding schema/xml template test to processor class.
